File 'client/game' splitted into new folder.
