+++ src/client/clientevent.h
@@ -22,6 +22,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include <string>
 #include "irrlichttypes_bloated.h"
 #include "hud.h"
+#include "nodedef.h"
 
 enum ClientEventType : u8
 {
