+++ util/travis/clang-format-whitelist.txt
@@ -25,12 +25,6 @@ src/clientmedia.cpp
 src/clientmedia.h
 src/clientobject.cpp
 src/clientobject.h
-src/client/render/core.cpp
-src/client/renderingengine.cpp
-src/client/render/interlaced.cpp
-src/client/render/plain.cpp
-src/client/render/sidebyside.cpp
-src/client/render/stereo.cpp
 src/client/tile.cpp
 src/client/tile.h
 src/client/fontengine.h
@@ -64,6 +58,26 @@ src/client/content_mapblock.h
 src/client/game.h
 src/client/mesh.h
 src/client/camera.cpp
+src/client/game/GameGlobalShaderConstantSetter.h
+src/client/game/TextDestPlayerInventory.h
+src/client/game/GameOnDemandSoundFetcher.h
+src/client/game/SoundMaker.h
+src/client/game/TextDestNodeMetadata.h
+src/client/game/Game.h
+src/client/game/Game.cpp
+src/client/game/PlayerInventoryFormSource.h
+src/client/game/NodeDugEvent.h
+src/client/game/LocalFormspecHandler.h
+src/client/game/NodeMetadataFormSource.h
+src/client/render/core.cpp
+src/client/render/plain.h
+src/client/render/factory.cpp
+src/client/render/core.cpp
+src/client/render/interlaced.cpp
+src/client/render/plain.cpp
+src/client/render/sidebyside.cpp
+src/client/render/stereo.cpp
+src/client/renderingengine.cpp
 src/client/sky.h
 src/client/mesh_generator_thread.cpp
 src/client/guiscalingfilter.h
